rewrite most of Bali.Reyong yet again, to support kotekan patterns
This time I use a table of absolute pitches to patterns, along with a
lot of functions to generate those tables, as opposed to the mixture of
tables and code that I had before.  Hopefully this will be able to account
for pattern variations and irregularities that will come later.

+++ Midi/Midi.hs
@@ -151,7 +151,7 @@ program_change bank program =
     [ ControlChange CC.bank msb, ControlChange CC.bank_lsb lsb
     , ProgramChange program
     ]
-    where (lsb, msb) = split14 (fromIntegral bank)
+    where (lsb, msb) = split14 bank
 
 -- | Emit a pitch bend sensitivity RPN message for the given range.
 pitch_bend_sensitivity :: Double -> [ChannelMessage]
